move resolve namespace factories to core
