+++ package.json
@@ -26,6 +26,7 @@
     "@types/csso": "^3.5.1",
     "@types/enhanced-resolve": "^3.0.6",
     "@types/express": "^4.17.2",
+    "@types/find-config": "^1.0.0",
     "@types/flat": "^0.0.28",
     "@types/jsdom": "^12.2.4",
     "@types/json-schema": "^7.0.3",
